requirements.txt
debugpy==1.4.1
Flask==2.0.2
gunicorn
h5py
# Keras==2.8.0
numpy==1.21.2
opencv-python-headless==4.6.0.66
protobuf==3.20.*
Pillow==7.1.0
pip==21.3.1
tensorflow-cpu==2.8.1
Werkzeug==2.1.2